Fix conversion warnings unrelated to shape (MATRIX TFVERTEX and DYNARRAY to fix later as very interconnected )

+++ src/libsac2c/flexsub/lubtree.c
@@ -101,7 +101,7 @@ LUBsetBlockIds (dynarray *eulertour, int blocksize)
                  */
 
                 if (prevdepth > currdepth) {
-                    blockid += pow (2, (blocksize - 2 - (j - (i + 1))));
+                    blockid += (int)pow (2.0,(double)(blocksize - 2 - (j - (i + 1))));
                 }
 
                 prevdepth = currdepth;
@@ -317,7 +317,7 @@ LUBcreatePartitions (dynarray *eulertour)
     if (totalelems == 1) {
         blocksize = 1;
     } else {
-        blocksize = log2 (totalelems) / 2.0;
+        blocksize = (int)(log2 (totalelems) / 2.0);
     }
 
     LUBINFO_BLOCKSIZE (lub) = blocksize;
@@ -482,13 +482,13 @@ LUBtreeLCAfromNodes (node *n1, node *n2, compinfo *ci)
 
         if (upperid / blocksize > lowerid / blocksize + 1) {
 
-            jump = floor (log2 (upperid / blocksize - lowerid / blocksize - 2));
+            jump = (int)floor (log2 (upperid / blocksize - lowerid / blocksize - 2));
 
             base = lowerid / blocksize + 1;
             e = DYNARRAY_ELEMS_POS (blockmin, getMatrixValue (intermat, base, jump));
             etindices[1] = *(int *)ELEM_DATA (e);
 
-            base = upperid / blocksize - 1 - pow (2, jump);
+            base = upperid / blocksize - 1 - (int)pow (2.0, (double)jump);
             e = DYNARRAY_ELEMS_POS (blockmin, getMatrixValue (intermat, base, jump));
             etindices[2] = *(int *)ELEM_DATA (e);
